+++ crates/cairo-lang-language-server/src/config.rs
@@ -9,9 +9,9 @@ use serde_json::Value;
 use tracing::{debug, error, warn};
 
 use crate::lsp::capabilities::client::ClientCapabilitiesExt;
-use crate::server::api::{LSPResult, LSPResultEx};
+use crate::lsp::result::{LSPResult, LSPResultEx};
 use crate::server::client::{Notifier, Requester};
-use crate::server::schedule::Task;
+use crate::server::schedule::task::Task;
 use crate::state::State;
 
 // TODO(mkaput): Write a macro that will auto-generate this struct and the `reload` logic.
