--- crates/cairo-lang-language-server/src/lsp/controller.rs
@@ -1,1 +1,0 @@
-
