+++ crates/cairo-lang-language-server/src/lsp/mod.rs
@@ -1,3 +1,3 @@
 pub(crate) mod capabilities;
-mod controller;
 pub mod ext;
+pub(crate) mod result;
