+++ crates/cairo-lang-language-server/src/server/schedule/mod.rs
@@ -2,22 +2,21 @@ use std::num::NonZeroUsize;
 
 use anyhow::Result;
 
-pub mod task;
-pub mod thread;
-
-pub use task::{BackgroundSchedule, Task};
+pub(crate) mod task;
+pub(crate) mod thread;
 
 use self::task::{BackgroundTaskBuilder, SyncTask};
 use self::thread::ThreadPriority;
 use super::client::Client;
 use crate::server::connection::ClientSender;
+use crate::server::schedule::task::{BackgroundSchedule, Task};
 use crate::state::State;
 
 /// The event loop thread is actually a secondary thread that we spawn from the
 /// _actual_ main thread. This secondary thread has a larger stack size
 /// than some OS defaults (Windows, for example) and is also designated as
 /// high-priority.
-pub(crate) fn event_loop_thread(
+pub fn event_loop_thread(
     func: impl FnOnce() -> Result<()> + Send + 'static,
 ) -> Result<thread::JoinHandle<Result<()>>> {
     // Override OS defaults to avoid stack overflows on platforms with low stack size defaults.
@@ -29,7 +28,7 @@ pub(crate) fn event_loop_thread(
         .spawn(func)?)
 }
 
-pub(crate) struct Scheduler<'s> {
+pub struct Scheduler<'s> {
     state: &'s mut State,
     client: Client<'s>,
     background_pool: thread::Pool,
@@ -47,7 +46,7 @@ impl<'s> Scheduler<'s> {
     /// Immediately sends a request of kind `R` to the client, with associated parameters.
     /// The task provided by `response_handler` will be dispatched as soon as the response
     /// comes back from the client.
-    pub(crate) fn request<R>(
+    pub fn request<R>(
         &mut self,
         params: R::Params,
         response_handler: impl Fn(R::Result) -> Task<'s> + 'static,
@@ -59,13 +58,13 @@ impl<'s> Scheduler<'s> {
     }
 
     /// Creates a task to handle a response from the client.
-    pub(crate) fn response(&mut self, response: lsp_server::Response) -> Task<'s> {
+    pub fn response(&mut self, response: lsp_server::Response) -> Task<'s> {
         self.client.requester.pop_response_task(response)
     }
 
     /// Dispatches a `task` by either running it as a blocking function or
     /// executing it on a background thread pool.
-    pub(crate) fn dispatch(&mut self, task: Task<'s>) {
+    pub fn dispatch(&mut self, task: Task<'s>) {
         match task {
             Task::Sync(SyncTask { func }) => {
                 let notifier = self.client.notifier();
